backend/requirements.txt
bleach==6.3.0
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
Deprecated==1.3.1
Flask==3.1.3
flask-cors==6.0.2
Flask-Limiter==4.1.1
Flask-SQLAlchemy==3.1.1
google_search_results==2.4.2
greenlet==3.3.2
gunicorn==25.3.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
limits==5.8.0
MarkupSafe==3.0.3
numpy==2.4.4
ordered-set==4.1.0
packaging==26.0
python-dotenv==1.2.2
requests==2.33.1
scikit-learn==1.8.0
scipy==1.17.1
SQLAlchemy==2.0.49
threadpoolctl==3.6.0
typing_extensions==4.15.0
urllib3==2.6.3
webencodings==0.5.1
Werkzeug==3.1.8
wrapt==2.1.2
